Moves tests to seperate files
