+++ server/src/api.rs
@@ -69,12 +69,9 @@ async fn get_filter_request(
 ) -> Result<impl Responder> {
     //TODO: construct filter
 
-    let tasks: Vec<task::Model> = task::Entity::find()
-        .all(data.as_ref())
-        .await
-        .map_err(|e| {
-            actix_web::error::ErrorInternalServerError(format!("couldn't filter tasks: {}", e))
-        })?;
+    let tasks: Vec<task::Model> = task::Entity::find().all(data.as_ref()).await.map_err(|e| {
+        actix_web::error::ErrorInternalServerError(format!("couldn't filter tasks: {}", e))
+    })?;
 
     Ok(web::Json(
         tasks.iter().map(|a| a.id).collect::<FilterResponse>(),
@@ -100,72 +97,4 @@ async fn create_task_request(
             actix_web::error::ErrorInternalServerError(format!("task not inserted {}", e))
         })?; //TODO handle this error better, for example for unique constraint violation
     Ok(web::Json(result_task.last_insert_id as CreateTaskResponse))
-}
-
-#[cfg(test)]
-mod tests {
-    use std::vec;
-
-    use actix_web::{dev::ServiceResponse, http::StatusCode};
-
-    use super::*;
-
-    #[actix_web::test]
-    async fn get_task_fails_with_bad_request() {
-        use actix_web::test;
-        use sea_orm::MockDatabase;
-
-        let db = MockDatabase::new(sea_orm::DatabaseBackend::Postgres);
-        let db_conn = db
-            .append_query_errors([sea_orm::error::DbErr::Query(
-                sea_orm::error::RuntimeErr::Internal("test".to_string()),
-            )])
-            .into_connection();
-        let db_data: web::Data<DatabaseConnection> = web::Data::new(db_conn);
-        let app = test::init_service(
-            actix_web::App::new()
-                .app_data(db_data)
-                .service(get_task_request),
-        )
-        .await;
-        let req = test::TestRequest::default()
-            .method(actix_web::http::Method::GET)
-            .set_json(ReadTaskShortRequest { task_id: 2 })
-            .uri("/task")
-            .to_request();
-        let resp: ServiceResponse = test::call_service(&app, req).await;
-        assert_eq!(resp.status(), StatusCode::INTERNAL_SERVER_ERROR);
-    }
-    #[actix_web::test]
-    async fn put_task_fails_with_bad_request() {
-        use actix_web::test;
-        use sea_orm::MockDatabase;
-        let db = MockDatabase::new(sea_orm::DatabaseBackend::Postgres);
-        let db_conn = db
-            .append_exec_errors([sea_orm::error::DbErr::Query(
-                sea_orm::error::RuntimeErr::Internal("test".to_string()),
-            )])
-            .append_query_errors([sea_orm::error::DbErr::Query(
-                sea_orm::error::RuntimeErr::Internal("test".to_string()),
-            )])
-            .into_connection();
-        let app = test::init_service(
-            actix_web::App::new()
-                .app_data(web::Data::new(db_conn))
-                .service(create_task_request),
-        )
-        .await;
-        let req = test::TestRequest::default()
-            .method(actix_web::http::Method::PUT)
-            .set_json(CreateTaskRequest {
-                name: "test".to_string(),
-                completed: false,
-                properties: vec![],
-                dependencies: vec![],
-            })
-            .uri("/task")
-            .to_request();
-        let resp: ServiceResponse = test::call_service(&app, req).await;
-        assert_eq!(resp.status(), StatusCode::INTERNAL_SERVER_ERROR);
-    }
 }
